+++ debugtools/formatter.py
@@ -3,6 +3,7 @@ An enhanced ``pprint.pformat`` that prints data structures in a readable HTML st
 """
 
 from django.core.exceptions import ObjectDoesNotExist, MultipleObjectsReturned
+from django.core.urlresolvers import NoReverseMatch
 from django.db import IntegrityError
 from django.db.models.base import Model
 from django.db.models.manager import Manager
@@ -250,7 +251,9 @@ def _format_lazy(value):
 def _try_call(func, extra_exceptions=()):
     try:
         return func()
-    except (TypeError, KeyError, AttributeError, ValueError, ObjectDoesNotExist, MultipleObjectsReturned, IntegrityError, AssertionError, NotImplementedError) as e:
+    except (TypeError, KeyError, AttributeError, ValueError,
+            ObjectDoesNotExist, MultipleObjectsReturned, IntegrityError, NoReverseMatch,
+            AssertionError, NotImplementedError) as e:
         return e
     except extra_exceptions as e:
         return e